pyproject.toml
[project]
name = "upstage-gangwon-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.122.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.29",
    "sqlalchemy-serializer>=1.4.22",
    "uvicorn[standard]>=0.38.0",
    "chromadb>=0.4.0",
    "openai==1.52.2",
    "httpx==0.27.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["test"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
